+++ internal/buildscripts/packaging/tests/deployments/salt/salt_test.py
@@ -112,7 +112,7 @@ def test_salt_with_fluentd(distro):
                 run_salt_apply(container, config)
                 verify_env_file(container)
                 assert wait_for(lambda: service_is_running(container))
-                if "opensuse" not in distro:
+                if "opensuse" not in distro and distro != "amazonlinux-2023":
                     assert container.exec_run("systemctl status td-agent").exit_code == 0
                 if collector_version == "latest":
                     verify_package_version(container, "splunk-otel-collector", collector_version, "0.34.0")
@@ -120,7 +120,7 @@ def test_salt_with_fluentd(distro):
                     verify_package_version(container, "splunk-otel-collector", collector_version)
         finally:
             run_container_cmd(container, f"journalctl -u {SERVICE_NAME} --no-pager")
-            if "opensuse" not in distro:
+            if "opensuse" not in distro and distro != "amazonlinux-2023":
                 run_container_cmd(container, "journalctl -u td-agent --no-pager")
                 if container.exec_run("test -f /var/log/td-agent/td-agent.log").exit_code == 0:
                     run_container_cmd(container, "cat /var/log/td-agent/td-agent.log")
@@ -211,8 +211,7 @@ def test_salt_with_instrumentation(distro):
                 run_salt_apply(container, config)
                 verify_env_file(container)
                 assert wait_for(lambda: service_is_running(container))
-                if "opensuse" not in distro:
-                    assert container.exec_run("systemctl status td-agent").exit_code == 0
+                assert container.exec_run("systemctl status td-agent").exit_code != 0
                 if version == "latest":
                     verify_package_version(container, "splunk-otel-auto-instrumentation", version, "0.48.0")
                 else:
@@ -220,10 +219,6 @@ def test_salt_with_instrumentation(distro):
                 verify_instrumentation_config(container)
         finally:
             run_container_cmd(container, f"journalctl -u {SERVICE_NAME} --no-pager")
-            if "opensuse" not in distro:
-                run_container_cmd(container, "journalctl -u td-agent --no-pager")
-                if container.exec_run("test -f /var/log/td-agent/td-agent.log").exit_code == 0:
-                    run_container_cmd(container, "cat /var/log/td-agent/td-agent.log")
 
 
 SERVICE_OWNER_CONFIG = f"""
@@ -234,6 +229,7 @@ splunk-otel-collector:
   splunk_api_url: '{SPLUNK_API_URL}'
   splunk_service_user: 'test-user'
   splunk_service_group: 'test-user'
+  install_fluentd: True
 """
 
 
@@ -256,13 +252,13 @@ def test_salt_service_owner(distro):
             assert wait_for(lambda: service_is_running(container, service_owner="test-user"))
             _, owner = run_container_cmd(container, f"stat -c '%U:%G' {SPLUNK_ENV_PATH}")
             assert owner.decode("utf-8").strip() == "test-user:test-user"
-            if "opensuse" not in distro:
+            if "opensuse" not in distro and distro != "amazonlinux-2023":
                 assert container.exec_run("systemctl status td-agent").exit_code == 0
                 _, owner = run_container_cmd(container, f"stat -c '%U:%G' {CONFIG_DIR}/fluentd/fluent.conf")
                 assert owner.decode("utf-8").strip() == "td-agent:td-agent"
         finally:
             run_container_cmd(container, f"journalctl -u {SERVICE_NAME} --no-pager")
-            if "opensuse" not in distro:
+            if "opensuse" not in distro and distro != "amazonlinux-2023":
                 run_container_cmd(container, "journalctl -u td-agent --no-pager")
                 if container.exec_run("test -f /var/log/td-agent/td-agent.log").exit_code == 0:
                     run_container_cmd(container, "cat /var/log/td-agent/td-agent.log")
@@ -272,7 +268,6 @@ CUSTOM_ENV_VARS_CONFIG = f"""
 splunk-otel-collector:
   splunk_access_token: '{SPLUNK_ACCESS_TOKEN}'
   splunk_realm: '{SPLUNK_REALM}'
-  install_fluentd: False
   collector_additional_env_vars:
     MY_CUSTOM_VAR1: value1
     MY_CUSTOM_VAR2: value2
@@ -298,5 +293,6 @@ def test_salt_custom_env_vars(distro):
             run_container_cmd(container, f"grep '^MY_CUSTOM_VAR1=value1$' {SPLUNK_ENV_PATH}")
             run_container_cmd(container, f"grep '^MY_CUSTOM_VAR2=value2$' {SPLUNK_ENV_PATH}")
             assert wait_for(lambda: service_is_running(container))
+            assert container.exec_run("systemctl status td-agent").exit_code != 0
         finally:
             run_container_cmd(container, f"journalctl -u {SERVICE_NAME} --no-pager")